feat: Qt6 migration - fix deprecated API usage across core, UI, packages, SWIG, and QML
- Replace QStringRef with QStringView, QRegExp with QRegularExpression, QMatrix with QTransform
- Update SWIG wrappers for Qt6 pointer-to-reference changes
- Fix QML QtQuick 2.15 import and mouse event signatures
- Remove Qt5 ChartDirector dependency from .pro
- Migrate timed-progressbar, linuxfb, and spine modules
- Update package translation string formats for compatibility

+++ src/ui/SpineEffectWidget.cpp
@@ -517,12 +517,12 @@ void SpineEffectWidget::renderSkeleton() {
         static int emptyLogCount = 0;
         if (++emptyLogCount <= 10)
             qWarning("[SpineEW] renderSkeleton: NO vertices generated! (batches=%d drawOrder=%d)",
-                     batches.size(), (int)drawOrder.size());
+                     (int)batches.size(), (int)drawOrder.size());
         return;
     }
 
     qWarning("[SpineEW] renderSkeleton: %d vertices, %d indices, %d batches (log#%d)",
-             _vertices.size(), _indices.size(), batches.size(), renderLogCount);
+             (int)_vertices.size(), (int)_indices.size(), (int)batches.size(), renderLogCount);
 
     // Only log render success for first 3 frames
     bool doRenderLog = (renderLogCount <= 3);
